pyproject.toml
[project]
name = "podcast-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-telegram-bot>=22.5",
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "xai-sdk>=1.4.1",
    "loguru>=0.7.3",
]
